Added functionality to retrieve friends list

+++ src/main/scala/twitter4s/entities/UserIds.scala
@@ -1,3 +1,3 @@
 package twitter4s.entities
 
-case class UserIds(ids: Seq[Long] = Seq.empty, next_cursor: Int, previous_cursor: Int)
+case class UserIds(ids: Seq[Long] = Seq.empty, next_cursor: Long, previous_cursor: Long)
