+++ src/main/scala/twitter4s/entities/UserIdsStringified.scala
@@ -1,3 +1,3 @@
 package twitter4s.entities
 
-case class UserIdsStringified(ids: Seq[String] = Seq.empty, next_cursor: Int, previous_cursor: Int)
+case class UserIdsStringified(ids: Seq[String] = Seq.empty, next_cursor: Long, previous_cursor: Long)
