+++ src/main/scala/twitter4s/http/clients/friendships/parameters/FollowingParameters.scala
@@ -4,6 +4,6 @@ import twitter4s.http.marshalling.Parameters
 
 case class FollowingParameters(user_id: Option[Long],
                                screen_name: Option[String],
-                               cursor: Int,
+                               cursor: Long,
                                count: Int,
                                stringify_ids: Boolean) extends Parameters
