+++ src/main/scala/twitter4s/http/clients/friendships/parameters/FriendshipParameters.scala
@@ -2,4 +2,4 @@ package twitter4s.http.clients.friendships.parameters
 
 import twitter4s.http.marshalling.Parameters
 
-case class FriendshipParameters(cursor: Int, stringify_ids: Boolean) extends Parameters
+case class FriendshipParameters(cursor: Long, stringify_ids: Boolean) extends Parameters
